Alter final screen's behaviour [SA-10194]
This changes the default Microsoft Teams app to redirect to /api/teams.php with two querystring params: the schoolbox domain and the meeting url. This allows us to return the meetingUrl's value to our own domain, and therefore inspect iframe contents to extract it and send it to the cke dialog.